refactor: move correctors to_8_5 package
These correctors were introduced individually on different versions. On
main they are added before releasing 8.5, so that's where this migration
task belongs.
