+++ engine/src/main/java/io/camunda/zeebe/engine/state/migration/to_8_5/ColumnFamilyPrefixCorrectionMigration.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.migration.to_8_4;
+package io.camunda.zeebe.engine.state.migration.to_8_5;
 
 import io.camunda.zeebe.engine.state.immutable.ProcessingState;
 import io.camunda.zeebe.engine.state.migration.MigrationTask;
