+++ engine/src/main/java/io/camunda/zeebe/engine/state/migration/to_8_5/corrections/ColumnFamily49Corrector.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.engine.state.migration.to_8_4.corrections;
+package io.camunda.zeebe.engine.state.migration.to_8_5.corrections;
 
 import io.camunda.zeebe.db.ColumnFamily;
 import io.camunda.zeebe.db.TransactionContext;
